feat(display): add 180 and 270 degree rotation

+++ main/bsp_display.h
@@ -61,6 +61,6 @@ uint8_t bsp_display_get_brightness(void);
  *   force_on: if true, always turn backlight on (used for SoftAP mode). */
 void bsp_display_apply_backlight_policy(bool force_on);
 
-/* Set LCD rotation in degrees (0, 90, 180, 270). Applies to hardware
- * immediately via ST7789 MADCTL. Must be re-applied after panel reset. */
-void bsp_display_set_rotation(uint8_t degrees);
+/* Set LCD rotation in degrees (0, 90, 180, 270). Applied by the display task
+ * via ST7789 MADCTL and re-applied after panel reset. */
+void bsp_display_set_rotation(uint16_t degrees);
